iot_socket_server/requirements.txt
click==8.1.3
cmake==3.26.4
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
dlib @ file:///C:/tem/16-06-2023/iot_socket_server/dlib-19.22.99-cp310-cp310-win_amd64.whl
dnspython==2.3.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fonttools==4.39.4
imutils==0.5.4
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
Pillow==9.5.0
pymongo==4.3.3
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
train==0.0.5
